requirements.txt
black
celery
Django
django-cors-headers
django-debug-toolbar
django-environ
django-extensions
django-redis
djangorestframework
django-import-export
ipdb
ipython
isort
psycopg2
redis

